% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/acompDS.R
\name{acompDS}
\alias{acompDS}
\title{Creates an Aitchison composition object for compositional data analysis}
\usage{
acompDS(
  X = NULL,
  parts = NULL,
  total = 1,
  warn.na = FALSE,
  detectionlimit = NULL,
  BDL = NULL,
  MAR = NULL,
  MNAR = NULL,
  SZ = NULL
)
}
\arguments{
\item{X}{name of the composition dataset or vector. Should be a character string.}

\item{parts}{vector containing the indices or names of the columns to be used.}

\item{total}{the total amount to be used, typically 1 or 100.}

\item{warn.na}{should the user be warned in case of NA, NaN or 0 coding different types of missing values?}

\item{detectionlimit}{a number, vector or matrix of positive numbers giving the detection limit.}

\item{BDL}{the code for 'Below Detection Limit' in X.}

\item{MAR}{the code for 'Missing At Random' in X.}

\item{MNAR}{the code for 'Missing Not At Random' in X.}

\item{SZ}{the code for 'Structural Zero' in X.}
}
\value{
an Aitchison composition object of class "acomp".
}
\description{
Creates an Aitchison composition object using the compositions::acomp() function
}
\details{
Serverside assign function \code{acompDS} called by clientside function
\code{ds.acomp}.
Returns an Aitchison composition object for compositional data analysis.
This function wraps the compositions::acomp() function for use in DataSHIELD.
For further details see help for \code{ds.acomp} function.
}
\author{
DataSHIELD Development Team
}
