% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ilrDS.R
\name{ilrDS}
\alias{ilrDS}
\title{Computes isometric log-ratio (ILR) coordinates for compositional data}
\usage{
ilrDS(X = NULL, V = NULL)
}
\arguments{
\item{X}{name of the composition dataset, matrix, data.frame or acomp object. Character string.}

\item{V}{an orthonormal basis matrix to be used for the ILR transform. If NULL, the default
basis used by compositions::ilr will be applied.}
}
\value{
a matrix-like object with ILR coordinates.
}
\description{
Creates ILR coordinates using the compositions::ilr() function on the server.
}
\details{
Serverside assign function \code{ilrDS} called by clientside function
\code{ds.ilr}. Returns ILR coordinates for compositional data analysis.
This function wraps the compositions::ilr() function for use in DataSHIELD.
For further details see help for \code{ds.ilr} function.
}
\author{
DataSHIELD Development Team
}
