% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lrEMDS.R
\name{lrEMDS}
\alias{lrEMDS}
\title{Creates a log-ratio EM algorithm object for compositional data imputation}
\usage{
lrEMDS(
  X = NULL,
  label = NULL,
  dl = NULL,
  rob = FALSE,
  ini.cov = "complete.obs",
  frac = 0.65,
  tolerance = 1e-04,
  max.iter = 50,
  rlm.maxit = 150,
  imp.missing = FALSE,
  suppress.print = FALSE,
  closure = NULL,
  z.warning = 0.8,
  z.delete = TRUE
)
}
\arguments{
\item{X}{name of the composition dataset or matrix. Should be a character string.}

\item{label}{unique label (numeric or character) used to denote unobserved values in X.}

\item{dl}{numeric vector or matrix of detection limits/thresholds.}

\item{rob}{logical value for robust estimation.}

\item{ini.cov}{initial estimation method.}

\item{frac}{fraction parameter for initial multiplicative simple replacement.}

\item{tolerance}{convergence criterion for the EM algorithm.}

\item{max.iter}{maximum number of iterations for the EM algorithm.}

\item{rlm.maxit}{maximum number of iterations for robust regression estimation.}

\item{imp.missing}{whether unobserved data are treated as missing data.}

\item{suppress.print}{suppress printed feedback.}

\item{closure}{closure value used to add a residual part if needed.}

\item{z.warning}{threshold for the proportion of modified zeros per part.}

\item{z.delete}{whether rows containing only zeros should be deleted.}
}
\value{
an imputed compositional data set (data.frame class).
}
\description{
Creates a log-ratio EM algorithm object using the zCompositions::lrEM() function
}
\details{
Serverside assign function \code{lrEMDS} called by clientside function
\code{ds.lrEM}.
Returns a log-ratio EM algorithm object for imputing left-censored data in compositional data sets.
This function wraps the zCompositions::lrEM() function for use in DataSHIELD.
For further details see help for \code{ds.lrEM} function.
}
\author{
DataSHIELD Development Team
}
