% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ilrDS.R
\name{ilrDS}
\alias{ilrDS}
\title{Computes isometric log-ratio (ILR) coordinates for compositional data}
\usage{
ilrDS(X = NULL, ...)
}
\arguments{
\item{X}{name of the composition dataset, matrix, data.frame or acomp object. Character string.}

\item{...}{additional arguments including flattened V matrix values and V_ncol parameter.
When V matrix is provided, it should be flattened and passed with V_ncol specifying the number of columns.}
}
\value{
a matrix-like object with ILR coordinates.
}
\description{
Creates ILR coordinates using the compositions::ilr() function on the server.
}
\details{
Serverside assign function \code{ilrDS} called by clientside function
\code{ds.ilr}. Returns ILR coordinates for compositional data analysis.
This function wraps the compositions::ilr() function for use in DataSHIELD.
For further details see help for \code{ds.ilr} function.
}
\author{
DataSHIELD Development Team
}
